refactor: separate image constants into constants/image.tsx
