+++ src/constants/index.ts
@@ -35,8 +35,3 @@ export const TABS = [ALBUMS, ARTISTS, TRACKS, PLAYLISTS] as const;
 export const DEFAULT_TAB = ALBUMS;
 
 export const ACCESS_TOKEN = 'access_token';
-
-export const PLACEHOLDER_IMAGE =
-  'https://placehold.co/600x600?text=NO IMAGE' as const;
-export const LOADING_IMAGE =
-  'https://placehold.co/600x600?text=LOADING' as const;
